+++ genno/tests/core/test_describe.py
@@ -2,6 +2,7 @@ import pandas as pd
 import pytest
 
 from genno import Computer, Quantity
+from genno.core.sparsedataarray import SparseDataArray
 
 
 @pytest.mark.usefixtures("parametrize_quantity_class")
@@ -9,7 +10,7 @@ def test_describe():
     c = Computer()
     c.add("foo", Quantity(pd.Series([42, 43])))
 
-    if Quantity.CLASS == "SparseDataArray":
+    if Quantity._get_class() is SparseDataArray:
         assert "\n".join(
             ["'foo':", "- <xarray.SparseDataArray (index: 2)>"]
         ) == c.describe("foo")