+++ genno/tests/core/test_quantity.py
@@ -1,12 +1,17 @@
 """Tests for genno.quantity."""
+import logging
 import re
 
 import pandas as pd
+import pint
 import pytest
 import xarray as xr
+from numpy import nan
 
 from genno import Computer, Quantity, computations
+from genno.core.attrseries import AttrSeries
 from genno.core.quantity import assert_quantity
+from genno.core.sparsedataarray import SparseDataArray
 from genno.testing import add_large_data, assert_qty_allclose, assert_qty_equal
 
 
@@ -19,12 +24,34 @@ class TestQuantity:
         da = xr.DataArray([0.8, 0.2], coords=[["oil", "water"]], dims=["p"])
         yield Quantity(da)
 
-    @pytest.mark.filterwarnings(
-        "ignore:.*default dtype for empty Series.*:DeprecationWarning"
+    @pytest.mark.parametrize(
+        "args, kwargs",
+        (
+            # Integer, converted to float() for sparse
+            ((3,), dict(units="kg")),
+            # Scalar object
+            ((object(),), dict(units="kg")),
+            # pd.Series
+            ((pd.Series([0, 1], index=["a", "b"], name="foo"),), dict(units="kg")),
+            # pd.DataFrame
+            (
+                (pd.DataFrame([[0], [1]], index=["a", "b"], columns=["foo"]),),
+                dict(units="kg"),
+            ),
+            pytest.param(
+                (
+                    pd.DataFrame(
+                        [[0, 1], [2, 3]], index=["a", "b"], columns=["foo", "bar"]
+                    ),
+                ),
+                dict(units="kg"),
+                marks=pytest.mark.xfail(raises=TypeError),
+            ),
+        ),
     )
-    def test_init(self):
+    def test_init(self, args, kwargs):
         """Instantiated from a scalar object."""
-        Quantity(object())
+        Quantity(*args, **kwargs)
 
     def test_assert(self, a):
         """Test assertions about Quantity.
@@ -77,9 +104,90 @@ class TestQuantity:
         a.attrs = {"bar": "foo"}
         assert_qty_equal(a, b, check_attrs=False)
 
-    def test_to_dataframe(self, a):
-        """Test Quantity.to_dataframe()."""
-        assert isinstance(a.to_dataframe(), pd.DataFrame)
+    @pytest.fixture()
+    def tri(self):
+        """Fixture returning triangular data to test fill, shift, etc."""
+        return Quantity(
+            xr.DataArray(
+                [
+                    [nan, nan, 1.0, nan, nan],
+                    [nan, 2, 3, 4, nan],
+                    [5, 6, 7, 8, 9],
+                ],
+                coords=[
+                    ("x", ["x0", "x1", "x2"]),
+                    ("y", ["y0", "y1", "y2", "y3", "y4"]),
+                ],
+            ),
+            units="kg",
+        )
+
+    def test_bfill(self, tri):
+        """Test Quantity.bfill()."""
+        if Quantity._get_class() is SparseDataArray:
+            pytest.xfail(reason="sparse.COO.flip() not implemented")
+
+        r1 = tri.bfill("x")
+        assert r1.loc["x0", "y0"] == tri.loc["x2", "y0"]
+
+        r2 = tri.bfill("y")
+        assert r2.loc["x0", "y0"] == tri.loc["x0", "y2"]
+
+    def test_copy_modify(self, a):
+        """Making a Quantity another produces a distinct attrs dictionary."""
+        assert 0 == len(a.attrs)
+
+        a.attrs["_unit"] = pint.Unit("km")
+
+        b = Quantity(a, units="kg")
+        assert pint.Unit("kg") == b.attrs["_unit"]
+
+        assert pint.Unit("km") == a.attrs["_unit"]
+
+    def test_cumprod(self, caplog, tri):
+        """Test Quantity.cumprod()."""
+        if Quantity._get_class() is SparseDataArray:
+            pytest.xfail(reason="sparse.COO.nancumprod() not implemented")
+
+        caplog.set_level(logging.INFO)
+
+        args = dict(axis=123) if Quantity._get_class() is AttrSeries else dict()
+        r1 = tri.cumprod("x", **args)
+        assert 1 * 3 * 7 == r1.loc["x2", "y2"]
+        if Quantity._get_class() is AttrSeries:
+            assert ["AttrSeries.cumprod(…, axis=…) is ignored"] == caplog.messages
+
+        r2 = tri.cumprod("y")
+        assert 2 * 3 == r2.loc["x1", "y2"]
+        assert 5 * 6 * 7 * 8 * 9 == r2.loc["x2", "y4"]
+
+    def test_ffill(self, tri):
+        """Test Quantity.ffill()."""
+
+        # Forward fill along "x" dimension results in no change
+        r1 = tri.ffill("x")
+        assert_qty_equal(tri, r1)
+
+        # Forward fill along y dimension works
+        r2 = tri.ffill("y")
+
+        # Check some filled values
+        assert r2.loc["x0", "y4"] == r2.loc["x0", "y3"] == tri.loc["x0", "y2"]
+
+    def test_shift(self, tri):
+        """Test Quantity.shift()."""
+        if Quantity._get_class() is SparseDataArray:
+            pytest.xfail(reason="sparse.COO.pad() not implemented")
+
+        r1 = tri.shift(x=1)
+        assert r1.loc["x2", "y1"] == tri.loc["x1", "y1"]
+
+        r2 = tri.shift(y=2)
+        assert r2.loc["x2", "y4"] == tri.loc["x2", "y2"]
+
+        with pytest.raises(NotImplementedError):
+            # AttrSeries only
+            tri.shift(x=1, y=2)
 
     def test_size(self):
         """Stress-test reporting of large, sparse quantities."""
@@ -100,3 +208,14 @@ class TestQuantity:
 
         # Result can be converted to pd.Series
         result.to_series()
+
+    def test_to_dataframe(self, a):
+        """Test Quantity.to_dataframe()."""
+        assert isinstance(a.to_dataframe(), pd.DataFrame)
+
+    def test_to_series(self, a):
+        """Test .to_series() on child classes, and Quantity.from_series."""
+        s = a.to_series()
+        assert isinstance(s, pd.Series)
+
+        Quantity.from_series(s)